rename cosocket to just socket
